chore: prevent running test against non-localhost db

+++ jest.setup.ts
@@ -74,6 +74,12 @@ type tableNames =
 async function truncateTables(prisma: PrismaClient, tableNames: tableNames[]) {
   const url = prisma['_engineConfig'].env.DATABASE_URL;
 
+  if(!url.include('localhost')) {
+    console.warn("wrong db url... please run the test using localhost")
+
+    return;
+  }
+
   for (const name of tableNames) {
     await prisma.$queryRawUnsafe(
       `TRUNCATE "${name}" RESTART IDENTITY CASCADE;`,
